Move and Rename logic
